+++ client/interledger/js/components/assets.js
@@ -2,36 +2,35 @@ import React from 'react';
 
 import AssetDetail from '../../../lib/js/react/components/asset_detail';
 
-const Assets = React.createClass({
 
-    propTypes: {
-        activeAccount: React.PropTypes.object,
-        assetListForAccount: React.PropTypes.array
-    },
+export default function Assets(props) {
+    const {
+        assetListForAccount
+    } = props;
 
-    render() {
-        const { assetListForAccount } = this.props;
-
-        if (assetListForAccount && assetListForAccount.length) {
-            return (
-                <div>
-                    {assetListForAccount
-                        .sort((a, b) => a.transaction.timestamp - b.transaction.timestamp)
-                        .map(asset => (
-                            <AssetDetail
-                                key={asset.id}
-                                asset={asset} />
-                        ))}
-                </div>
-            );
-        } else {
-            return (
-                <div className="content-text">
-                    No assets found in BigchainDB. Start trading...
-                </div>
-            );
-        }
+    if (assetListForAccount && assetListForAccount.length) {
+        return (
+            <div>
+                {assetListForAccount
+                    .sort((a, b) => a.transaction.timestamp - b.transaction.timestamp)
+                    .map(asset => (
+                        <AssetDetail
+                            key={asset.id}
+                            asset={asset} />
+                    ))}
+            </div>
+        );
+    } else {
+        return (
+            <div className="content-text">
+                No assets found in BigchainDB. Start trading...
+            </div>
+        );
     }
-});
+}
 
-export default Assets;
+
+Assets.propTypes = {
+    activeAccount: React.PropTypes.object,
+    assetListForAccount: React.PropTypes.array
+};
