+++ client/on_the_record/js/components/on_the_record.js
@@ -49,7 +49,11 @@ const OnTheRecord = React.createClass({
     },
 
     render() {
-        const { activeAccount, assetList, assetMeta } = this.state;
+        const {
+            activeAccount,
+            assetList,
+            assetMeta
+        } = this.state;
 
         return (
             <div>
